catkin plugin: extract rosdep into pullers package
This is paving the way for support of Ament, ROS 2's build system, which
also uses rosdep.

Resolve #1464.
