+++ snapcraft/internal/pluginhandler/__init__.py
@@ -144,6 +144,9 @@ class PluginHandler:
                     'unknown plugin: {}'.format(plugin_name))
 
         plugin = _get_plugin(module)
+        if not plugin:
+            raise errors.PluginError(
+                'no plugin found in module: {}'.format(plugin_name))
         _validate_pull_and_build_properties(
             plugin_name, plugin, part_schema, definitions_schema)
         options = _make_options(
