+++ snapcraft/plugins/catkin.py
@@ -1,6 +1,6 @@
 # -*- Mode:Python; indent-tabs-mode:nil; tab-width:4 -*-
 #
-# Copyright (C) 2015-2016 Canonical Ltd
+# Copyright (C) 2015-2017 Canonical Ltd
 #
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License version 3 as
@@ -64,6 +64,7 @@ import subprocess
 import textwrap
 
 import snapcraft
+from snapcraft.plugins.internal import pullers
 from snapcraft import (
     common,
     file_utils,
@@ -332,9 +333,13 @@ deb http://${{security}}.ubuntu.com/${{suffix}} {0}-security main universe
         compilers.setup()
 
         # Use rosdep for dependency detection and resolution
-        rosdep = _Rosdep(self.options.rosdistro, self._ros_package_path,
-                         self._rosdep_path, self.PLUGIN_STAGE_SOURCES,
-                         self.project)
+        rosdep = pullers.rosdep.Rosdep(
+            ros_distro=self.options.rosdistro,
+            ros_package_path=self._ros_package_path,
+            rosdep_path=self._rosdep_path,
+            ubuntu_distro=_ROS_RELEASE_MAP[self.options.rosdistro],
+            ubuntu_sources=self.PLUGIN_STAGE_SOURCES,
+            project=self.project)
         rosdep.setup()
 
         self._setup_dependencies(rosdep, catkin)
@@ -665,7 +670,7 @@ def _find_system_dependencies(catkin_packages, rosdep, catkin):
             # it.
             try:
                 these_dependencies = rosdep.resolve_dependency(dependency)
-            except SystemDependencyNotFoundError:
+            except pullers.rosdep.RosdepDependencyNotFoundError:
                 raise RuntimeError(
                     "Package {!r} isn't a valid system dependency. "
                     "Did you forget to add it to catkin-packages? If "
@@ -691,130 +696,11 @@ def _handle_rosinstall_files(wstool, source_path, rosinstall_files):
     wstool.update()
 
 
-class SystemDependencyNotFoundError(errors.SnapcraftError):
-    fmt = '{system_dependency!r} does not resolve to a system dependency'
-
-    def __init__(self, system_dependency):
-        super().__init__(system_dependency=system_dependency)
-
-
 class CatkinPackageNotFoundError(errors.SnapcraftError):
     fmt = 'Unable to find Catkin package {package_name!r}'
 
     def __init__(self, package_name):
         super().__init__(package_name=package_name)
-
-
-class _Rosdep:
-    def __init__(self, ros_distro, ros_package_path, rosdep_path,
-                 ubuntu_sources, project):
-        self._ros_distro = ros_distro
-        self._ros_package_path = ros_package_path
-        self._ubuntu_sources = ubuntu_sources
-        self._rosdep_path = rosdep_path
-        self._rosdep_install_path = os.path.join(self._rosdep_path, 'install')
-        self._rosdep_sources_path = os.path.join(self._rosdep_path,
-                                                 'sources.list.d')
-        self._rosdep_cache_path = os.path.join(self._rosdep_path, 'cache')
-        self._project = project
-
-    def setup(self):
-        # Make sure we can run multiple times without error, while leaving the
-        # capability to re-initialize, by making sure we clear the sources.
-        if os.path.exists(self._rosdep_sources_path):
-            shutil.rmtree(self._rosdep_sources_path)
-
-        os.makedirs(self._rosdep_sources_path)
-        os.makedirs(self._rosdep_install_path, exist_ok=True)
-        os.makedirs(self._rosdep_cache_path, exist_ok=True)
-
-        # rosdep isn't necessarily a dependency of the project, and we don't
-        # want to bloat the .snap more than necessary. So we'll unpack it
-        # somewhere else, and use it from there.
-        logger.info('Preparing to fetch rosdep...')
-        ubuntu = repo.Ubuntu(self._rosdep_path, sources=self._ubuntu_sources,
-                             project_options=self._project)
-
-        logger.info('Fetching rosdep...')
-        ubuntu.get(['python-rosdep'])
-
-        logger.info('Installing rosdep...')
-        ubuntu.unpack(self._rosdep_install_path)
-
-        logger.info('Initializing rosdep database...')
-        try:
-            self._run(['init'])
-        except subprocess.CalledProcessError as e:
-            output = e.output.decode('utf8').strip()
-            raise RuntimeError(
-                'Error initializing rosdep database:\n{}'.format(output))
-
-        logger.info('Updating rosdep database...')
-        try:
-            self._run(['update'])
-        except subprocess.CalledProcessError as e:
-            output = e.output.decode('utf8').strip()
-            raise RuntimeError(
-                'Error updating rosdep database:\n{}'.format(output))
-
-    def get_dependencies(self, package_name):
-        try:
-            output = self._run(['keys', package_name]).strip()
-            if output:
-                return output.split('\n')
-            else:
-                return []
-        except subprocess.CalledProcessError:
-            raise FileNotFoundError(
-                'Unable to find Catkin package "{}"'.format(package_name))
-
-    def resolve_dependency(self, dependency_name):
-        try:
-            # rosdep needs three pieces of information here:
-            #
-            # 1) The dependency we're trying to lookup.
-            # 2) The rosdistro being used.
-            # 3) The version of Ubuntu being used. We're telling rosdep to
-            #    resolve dependencies using the version of Ubuntu that
-            #    corresponds to the ROS release (even if we're running on
-            #    something else).
-            output = self._run(['resolve', dependency_name, '--rosdistro',
-                                self._ros_distro, '--os',
-                                'ubuntu:{}'.format(
-                                    _ROS_RELEASE_MAP[self._ros_distro])])
-        except subprocess.CalledProcessError:
-            raise SystemDependencyNotFoundError(dependency_name)
-
-        # Everything that isn't a package name is prepended with the pound
-        # sign, so we'll ignore everything with that.
-        delimiters = re.compile(r'\n|\s')
-        lines = delimiters.split(output)
-        return [line for line in lines if not line.startswith('#')]
-
-    def _run(self, arguments):
-        env = os.environ.copy()
-
-        # We want to make sure we use our own rosdep (which is python)
-        env['PATH'] = os.path.join(self._rosdep_install_path, 'usr', 'bin')
-        env['PYTHONPATH'] = os.path.join(self._rosdep_install_path, 'usr',
-                                         'lib', 'python2.7', 'dist-packages')
-
-        # By default, rosdep uses /etc/ros/rosdep to hold its sources list. We
-        # don't want that here since we don't want to touch the host machine
-        # (not to mention it would require sudo), so we can redirect it via
-        # this environment variable
-        env['ROSDEP_SOURCE_PATH'] = self._rosdep_sources_path
-
-        # By default, rosdep saves its cache in $HOME/.ros, which we shouldn't
-        # access here, so we'll redirect it with this environment variable.
-        env['ROS_HOME'] = self._rosdep_cache_path
-
-        # This environment variable tells rosdep which directory to recursively
-        # search for packages.
-        env['ROS_PACKAGE_PATH'] = self._ros_package_path
-
-        return subprocess.check_output(['rosdep'] + arguments,
-                                       env=env).decode('utf8').strip()
 
 
 class _Compilers:
