+++ snapcraft/tests/plugins/test_catkin.py
@@ -32,6 +32,7 @@ from testtools.matchers import (
 
 import snapcraft
 from snapcraft.plugins import catkin
+from snapcraft.plugins.internal import pullers
 from snapcraft import (
     repo,
     tests,
@@ -65,6 +66,7 @@ class CatkinPluginBaseTestCase(tests.TestCase):
 
         class props:
             rosdistro = 'indigo'
+            ubuntu_distro = 'trusty'
             catkin_packages = ['my_package']
             source_space = 'src'
             source_subdir = None
@@ -84,7 +86,8 @@ class CatkinPluginBaseTestCase(tests.TestCase):
         self.dependencies_mock = patcher.start()
         self.addCleanup(patcher.stop)
 
-        patcher = mock.patch('snapcraft.plugins.catkin._Rosdep')
+        patcher = mock.patch(
+            'snapcraft.plugins.internal.pullers.rosdep.Rosdep')
         self.rosdep_mock = patcher.start()
         self.addCleanup(patcher.stop)
 
@@ -97,10 +100,15 @@ class CatkinPluginBaseTestCase(tests.TestCase):
         self.addCleanup(patcher.stop)
 
     def assert_rosdep_setup(self, rosdistro, package_path, rosdep_path,
-                            sources):
+                            ubuntu_distro, sources):
         self.rosdep_mock.assert_has_calls([
-            mock.call(rosdistro, package_path, rosdep_path, sources,
-                      self.project_options),
+            mock.call(
+                ros_distro=rosdistro,
+                ros_package_path=package_path,
+                rosdep_path=rosdep_path,
+                ubuntu_distro=ubuntu_distro,
+                ubuntu_sources=sources,
+                project=self.project_options),
             mock.call().setup()])
 
     def assert_wstool_setup(self, package_path, wstool_path, sources):
@@ -794,6 +802,7 @@ class PullTestCase(CatkinPluginBaseTestCase):
             self.properties.rosdistro,
             os.path.join(plugin.sourcedir, 'src'),
             os.path.join(plugin.partdir, 'rosdep'),
+            self.properties.ubuntu_distro,
             plugin.PLUGIN_STAGE_SOURCES)
 
         self.wstool_mock.assert_not_called()
@@ -835,6 +844,7 @@ class PullTestCase(CatkinPluginBaseTestCase):
             self.properties.rosdistro,
             os.path.join(plugin.sourcedir, 'src'),
             os.path.join(plugin.partdir, 'rosdep'),
+            self.properties.ubuntu_distro,
             plugin.PLUGIN_STAGE_SOURCES)
 
         self.wstool_mock.assert_not_called()
@@ -879,6 +889,7 @@ class PullTestCase(CatkinPluginBaseTestCase):
             self.properties.rosdistro,
             os.path.join(plugin.sourcedir, 'src'),
             os.path.join(plugin.partdir, 'rosdep'),
+            self.properties.ubuntu_distro,
             plugin.PLUGIN_STAGE_SOURCES)
 
         self.wstool_mock.assert_not_called()
@@ -912,6 +923,7 @@ class PullTestCase(CatkinPluginBaseTestCase):
             self.properties.rosdistro,
             os.path.join(plugin.sourcedir, 'src'),
             os.path.join(plugin.partdir, 'rosdep'),
+            self.properties.ubuntu_distro,
             plugin.PLUGIN_STAGE_SOURCES)
 
         self.assert_wstool_setup(
@@ -1118,7 +1130,7 @@ class FindSystemDependenciesTestCase(tests.TestCase):
     def test_find_system_dependencies_missing_local_dependency(self):
         # Setup a dependency on a non-existing package, and it doesn't resolve
         # to a system dependency.'
-        exception = catkin.SystemDependencyNotFoundError('foo')
+        exception = pullers.rosdep.RosdepDependencyNotFoundError('foo')
         self.rosdep_mock.resolve_dependency.side_effect = exception
 
         raised = self.assertRaises(
@@ -1155,157 +1167,6 @@ class HandleRosinstallFilesTestCase(tests.TestCase):
             mock.call(os.path.join('source_path', 'file1')),
             mock.call(os.path.join('source_path', 'file2'))
         ])
-
-
-class RosdepTestCase(tests.TestCase):
-
-    def setUp(self):
-        super().setUp()
-        self.project = snapcraft.ProjectOptions()
-        self.rosdep = catkin._Rosdep('kinetic', 'package_path',
-                                     'rosdep_path', 'sources',
-                                     self.project)
-
-        patcher = mock.patch('snapcraft.repo.Ubuntu')
-        self.ubuntu_mock = patcher.start()
-        self.addCleanup(patcher.stop)
-
-        patcher = mock.patch('subprocess.check_output')
-        self.check_output_mock = patcher.start()
-        self.addCleanup(patcher.stop)
-
-    def test_setup(self):
-        # Return something other than a Mock to ease later assertions
-        self.check_output_mock.return_value = b''
-
-        self.rosdep.setup()
-
-        # Verify that only rosdep was installed (no other .debs)
-        self.assertEqual(self.ubuntu_mock.call_count, 1)
-        self.assertEqual(self.ubuntu_mock.return_value.get.call_count, 1)
-        self.assertEqual(self.ubuntu_mock.return_value.unpack.call_count, 1)
-        self.ubuntu_mock.assert_has_calls([
-            mock.call(self.rosdep._rosdep_path, sources='sources',
-                      project_options=self.project),
-            mock.call().get(['python-rosdep']),
-            mock.call().unpack(self.rosdep._rosdep_install_path)])
-
-        # Verify that rosdep was initialized and updated
-        self.assertEqual(self.check_output_mock.call_count, 2)
-        self.check_output_mock.assert_has_calls([
-            mock.call(['rosdep', 'init'], env=mock.ANY),
-            mock.call(['rosdep', 'update'], env=mock.ANY)
-        ])
-
-    def test_setup_can_run_multiple_times(self):
-        self.rosdep.setup()
-
-        # Make sure running setup() again doesn't have problems with the old
-        # environment
-        # An exception will be raised if setup can't be called twice.
-        self.rosdep.setup()
-
-    def test_setup_initialization_failure(self):
-        def run(args, **kwargs):
-            if args == ['rosdep', 'init']:
-                raise subprocess.CalledProcessError(1, 'foo', b'bar')
-
-        self.check_output_mock.side_effect = run
-
-        raised = self.assertRaises(RuntimeError, self.rosdep.setup)
-
-        self.assertEqual(str(raised),
-                         'Error initializing rosdep database:\nbar')
-
-    def test_setup_update_failure(self):
-        def run(args, **kwargs):
-            if args == ['rosdep', 'update']:
-                raise subprocess.CalledProcessError(1, 'foo', b'bar')
-
-            return mock.DEFAULT
-
-        self.check_output_mock.side_effect = run
-
-        raised = self.assertRaises(RuntimeError, self.rosdep.setup)
-
-        self.assertEqual(str(raised),
-                         'Error updating rosdep database:\nbar')
-
-    def test_get_dependencies(self):
-        self.check_output_mock.return_value = b'foo\nbar\nbaz'
-
-        self.assertEqual(self.rosdep.get_dependencies('foo'),
-                         ['foo', 'bar', 'baz'])
-
-        self.check_output_mock.assert_called_with(['rosdep', 'keys', 'foo'],
-                                                  env=mock.ANY)
-
-    def test_get_dependencies_no_dependencies(self):
-        self.check_output_mock.return_value = b''
-
-        self.assertEqual(self.rosdep.get_dependencies('foo'), [])
-
-    def test_get_dependencies_invalid_package(self):
-        self.check_output_mock.side_effect = subprocess.CalledProcessError(
-            1, 'foo')
-
-        raised = self.assertRaises(
-            FileNotFoundError,
-            self.rosdep.get_dependencies, 'bar')
-
-        self.assertEqual(str(raised),
-                         'Unable to find Catkin package "bar"')
-
-    def test_resolve_dependency(self):
-        self.check_output_mock.return_value = b'#apt\nmylib-dev'
-
-        self.assertEqual(self.rosdep.resolve_dependency('foo'), ['mylib-dev'])
-
-        self.check_output_mock.assert_called_with(
-            ['rosdep', 'resolve', 'foo', '--rosdistro', 'kinetic', '--os',
-             'ubuntu:xenial'],
-            env=mock.ANY)
-
-    def test_resolve_invalid_dependency(self):
-        self.check_output_mock.side_effect = subprocess.CalledProcessError(
-            1, 'foo')
-
-        raised = self.assertRaises(
-            catkin.SystemDependencyNotFoundError,
-            self.rosdep.resolve_dependency, 'bar')
-
-        self.assertEqual(str(raised),
-                         "'bar' does not resolve to a system dependency")
-
-    def test_resolve_no_dependency(self):
-        self.check_output_mock.return_value = b'#apt'
-
-        self.assertEqual(self.rosdep.resolve_dependency('bar'), [])
-
-    def test_resolve_multiple_dependencies(self):
-        self.check_output_mock.return_value = b'#apt\nlib1 lib2'
-
-        self.assertEqual(self.rosdep.resolve_dependency('foo'),
-                         ['lib1', 'lib2'])
-
-    def test_run(self):
-        rosdep = self.rosdep
-        rosdep._run(['qux'])
-
-        class check_env():
-            def __eq__(self, env):
-                rosdep_sources_path = rosdep._rosdep_sources_path
-                return (
-                    env['PATH'] == os.path.join(rosdep._rosdep_install_path,
-                                                'usr', 'bin') and
-                    env['PYTHONPATH'] == os.path.join(
-                        rosdep._rosdep_install_path, 'usr', 'lib', 'python2.7',
-                        'dist-packages') and
-                    env['ROSDEP_SOURCE_PATH'] == rosdep_sources_path and
-                    env['ROS_HOME'] == rosdep._rosdep_cache_path and
-                    env['ROS_PACKAGE_PATH'] == rosdep._ros_package_path)
-
-        self.check_output_mock.assert_called_with(mock.ANY, env=check_env())
 
 
 class CompilersTestCase(tests.TestCase):
